feat: Initialize app language based on browser
This change updates the application to detect the user's browser language upon initial load.
The default language is now set to French ('fr') if the browser language is detected as French, otherwise it defaults to English ('en').
This setting is also used to initialize the language from localStorage if a preference has been previously saved, providing a more personalized user experience.